Add option to limit mapping to api_attributes_of_interest (#5)
This change allows developers to reduce unnecessary work performed in attribute map in scenarios where only a select set of attributes needs to be mapped. Such attributes can be indicated by an optional `api_attributes_of_interest` passed to the transformer via its options hash.

This change also fixes a bug in the abstract transformer that raised unknown attribute errors with an incorrect class. Instead, the transformer will raise an internal error with a custom error message that includes the attribute name and correct class name.

---------

Co-authored-by: Matt Farmer <mrfarmer777@gmail.com>